Removed time-zone name support from ISODateTimeParser
There is an inconsistency in time zone name parsing between Java 8
and later versions. While Java 8 sets the time to the local time in
the specified time zone, later versions apply the zone id to the
time in GMT. As a result, a string like "10:00:00[America/New_York]"
is parsed to the same time as "15:00:00Z" or "10:00:00-05:00" in
Java 8, while later versions parse the same string to the same time
as "10:00:00Z" or "05:00:00-05:00"

Since zone name parsing is not needed for compatibility with Joda's
ISODateTimeFormat.dateTimeParser, it can be removed entirely.